Add tests rootless run
The options here were either to override the EUID in the code, or require the E2E tests to be run one as sudo, and one as not sudo.
The latter was chosen as the changes to the code for tests were somewhat heavy.

Add sudo test run to the actions.

Signed-off-by: luke <luke.parkin@arm.com>